Handle highlightDiff error by falling back to plain diff.
In case the highlighting algorithm encounters an error on some diff,
we don't want to abort entirely. Fall back to showing plain diff and
log the error for investigation.

This is an intermediate solution until the highlighting algorithm is
improved to handle the problematic case (which may take longer).

Helps https://dmitri.shuralyov.com/app/changes/...$issues/2.