mark changes as read
The responsibility of marking changes as read is being moved from
the service level to the changes application. It has already been
removed from the change service. In this commit, it's being added
to the changes application. Use the existing Options.Notifications
field that provides a notification service and its MarkRead method
for accomplishing this.

In general, it's easier for the changes application to know (with
higher certainty) when a change has been read by a user. For now,
it continues to happen whenever the HTTP handler serves a page that
displays a change. In the future, we can decide to mark a thread as
read when the user scrolls down and sees the new content, instead of
right away on page load.

Also update cmd/changesdev for githubapi.NewService API change.

Follows https://dmitri.shuralyov.com/service/change/githubapi$commit/09fc1cdc80ff92256db38ada7ae5bdecad9db138